Done separating the mini drawer in two components called SideList and Dashboard
