+++ 02. MERN - CouchSurfingApp | MERN | Firebase | Mapbox | JWT | Google Auth/place-to-stay/client/src/pages/dashboard/SideList.js
@@ -1,0 +1,190 @@
+import React from "react";
+
+import MuiDrawer from "@mui/material/Drawer";
+import {
+	Box,
+	Divider,
+	IconButton,
+	List,
+	ListItem,
+	ListItemButton,
+	ListItemIcon,
+	ListItemText,
+	Typography,
+	styled,
+} from "@mui/material";
+import { ChevronLeft, Inbox, Mail } from "@mui/icons-material";
+
+const drawerWidth = 240;
+
+const openedMixin = (theme) => ({
+	width: drawerWidth,
+	transition: theme.transitions.create("width", {
+		easing: theme.transitions.easing.sharp,
+		duration: theme.transitions.duration.enteringScreen,
+	}),
+	overflowX: "hidden",
+});
+
+const closedMixin = (theme) => ({
+	transition: theme.transitions.create("width", {
+		easing: theme.transitions.easing.sharp,
+		duration: theme.transitions.duration.leavingScreen,
+	}),
+	overflowX: "hidden",
+	width: `calc(${theme.spacing(7)} + 1px)`,
+	[theme.breakpoints.up("sm")]: {
+		width: `calc(${theme.spacing(8)} + 1px)`,
+	},
+});
+
+const DrawerHeader = styled("div")(({ theme }) => ({
+	display: "flex",
+	alignItems: "center",
+	justifyContent: "flex-end",
+	padding: theme.spacing(0, 1),
+	// necessary for content to be below app bar
+	...theme.mixins.toolbar,
+}));
+
+const Drawer = styled(MuiDrawer, {
+	shouldForwardProp: (prop) => prop !== "open",
+})(({ theme, open }) => ({
+	width: drawerWidth,
+	flexShrink: 0,
+	whiteSpace: "nowrap",
+	boxSizing: "border-box",
+	...(open && {
+		...openedMixin(theme),
+		"& .MuiDrawer-paper": openedMixin(theme),
+	}),
+	...(!open && {
+		...closedMixin(theme),
+		"& .MuiDrawer-paper": closedMixin(theme),
+	}),
+}));
+
+const SideList = ({ open, setOpen }) => {
+	return (
+		<>
+			<Drawer variant="permanent" open={open}>
+				<DrawerHeader>
+					{/* This closes the drawer */}
+					<IconButton
+						onClick={() => {
+							setOpen(false);
+						}}
+					>
+						<ChevronLeft />
+					</IconButton>
+				</DrawerHeader>
+				<Divider />
+				<List>
+					{["Inbox", "Starred", "Send email", "Drafts"].map(
+						(text, index) => (
+							<ListItem
+								key={text}
+								disablePadding
+								sx={{ display: "block" }}
+							>
+								<ListItemButton
+									sx={{
+										minHeight: 48,
+										justifyContent: open
+											? "initial"
+											: "center",
+										px: 2.5,
+									}}
+								>
+									<ListItemIcon
+										sx={{
+											minWidth: 0,
+											mr: open ? 3 : "auto",
+											justifyContent: "center",
+										}}
+									>
+										{index % 2 === 0 ? <Inbox /> : <Mail />}
+									</ListItemIcon>
+									<ListItemText
+										primary={text}
+										sx={{ opacity: open ? 1 : 0 }}
+									/>
+								</ListItemButton>
+							</ListItem>
+						)
+					)}
+				</List>
+				<Divider />
+				<List>
+					{["All mail", "Trash", "Spam"].map((text, index) => (
+						<ListItem
+							key={text}
+							disablePadding
+							sx={{ display: "block" }}
+						>
+							<ListItemButton
+								sx={{
+									minHeight: 48,
+									justifyContent: open ? "initial" : "center",
+									px: 2.5,
+								}}
+							>
+								<ListItemIcon
+									sx={{
+										minWidth: 0,
+										mr: open ? 3 : "auto",
+										justifyContent: "center",
+									}}
+								>
+									{index % 2 === 0 ? <Inbox /> : <Mail />}
+								</ListItemIcon>
+								<ListItemText
+									primary={text}
+									sx={{ opacity: open ? 1 : 0 }}
+								/>
+							</ListItemButton>
+						</ListItem>
+					))}
+				</List>
+			</Drawer>
+			<Box component="main" sx={{ flexGrow: 1, p: 3 }}>
+				<DrawerHeader />
+				<Typography paragraph>
+					Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed
+					do eiusmod tempor incididunt ut labore et dolore magna
+					aliqua. Rhoncus dolor purus non enim praesent elementum
+					facilisis leo vel. Risus at ultrices mi tempus imperdiet.
+					Semper risus in hendrerit gravida rutrum quisque non tellus.
+					Convallis convallis tellus id interdum velit laoreet id
+					donec ultrices. Odio morbi quis commodo odio aenean sed
+					adipiscing. Amet nisl suscipit adipiscing bibendum est
+					ultricies integer quis. Cursus euismod quis viverra nibh
+					cras. Metus vulputate eu scelerisque felis imperdiet proin
+					fermentum leo. Mauris commodo quis imperdiet massa
+					tincidunt. Cras tincidunt lobortis feugiat vivamus at augue.
+					At augue eget arcu dictum varius duis at consectetur lorem.
+					Velit sed ullamcorper morbi tincidunt. Lorem donec massa
+					sapien faucibus et molestie ac.
+				</Typography>
+				<Typography paragraph>
+					Consequat mauris nunc congue nisi vitae suscipit. Fringilla
+					est ullamcorper eget nulla facilisi etiam dignissim diam.
+					Pulvinar elementum integer enim neque volutpat ac tincidunt.
+					Ornare suspendisse sed nisi lacus sed viverra tellus. Purus
+					sit amet volutpat consequat mauris. Elementum eu facilisis
+					sed odio morbi. Euismod lacinia at quis risus sed vulputate
+					odio. Morbi tincidunt ornare massa eget egestas purus
+					viverra accumsan in. In hendrerit gravida rutrum quisque non
+					tellus orci ac. Pellentesque nec nam aliquam sem et tortor.
+					Habitant morbi tristique senectus et. Adipiscing elit duis
+					tristique sollicitudin nibh sit. Ornare aenean euismod
+					elementum nisi quis eleifend. Commodo viverra maecenas
+					accumsan lacus vel facilisis. Nulla posuere sollicitudin
+					aliquam ultrices sagittis orci a.
+				</Typography>
+			</Box>
+		</>
+	);
+};
+
+export default SideList;
